+++ lib/src/button/sbb_secondary_button.dart
@@ -15,7 +15,7 @@ import '../../design_system_flutter.dart';
 /// * <https://digital.sbb.ch/de/design-system-mobile-new/elemente/button>
 class SBBSecondaryButton extends StatelessWidget {
   const SBBSecondaryButton({
-    super. key,
+    super.key,
     required this.label,
     this.isLoading = false,
     required this.onPressed,
@@ -30,19 +30,14 @@ class SBBSecondaryButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final style = SBBBaseStyle.of(context);
-    final isWeb = style.hostPlatform == HostPlatform.web;
     final buttonStyles = SBBButtonStyles.of(context);
     return OutlinedButton(
-      style: isWeb
-          ? buttonStyles.secondaryWebLean
-          : Theme.of(context).outlinedButtonTheme.style?.copyWith(
-                // workaround for web
-                padding: SBBTheme.allStates(EdgeInsets.zero),
-              ),
+      style: Theme.of(context).outlinedButtonTheme.style?.copyWith(
+            padding: SBBTheme.allStates(EdgeInsets.zero),
+          ),
       onPressed: isLoading ? null : onPressed,
       focusNode: focusNode,
       child: Padding(
-        // workaround for web
         padding: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -55,37 +50,6 @@ class SBBSecondaryButton extends StatelessWidget {
             buttonStyles.buttonLabelBuilder!(context, label),
           ],
         ),
-      ),
-    );
-  }
-}
-
-class SBBGhostButton extends StatelessWidget {
-  const SBBGhostButton({
-    Key? key,
-    required this.label,
-    required this.onPressed,
-    this.focusNode,
-  }) : super(key: key);
-
-  final String label;
-  final VoidCallback? onPressed;
-  final FocusNode? focusNode;
-
-  @override
-  Widget build(BuildContext context) {
-    if (SBBBaseStyle.of(context).hostPlatform == HostPlatform.native)
-      debugPrint('WARNING: Ghost button should only be used for web platform.');
-    final styles = SBBButtonStyles.of(context);
-    return OutlinedButton(
-      style: styles.ghostWebLean,
-      onPressed: onPressed,
-      focusNode: focusNode,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          styles.buttonLabelBuilder!(context, label),
-        ],
       ),
     );
   }